Changed Movement Physics
--PLAYER CHANGES
Replaced Jumping variable with grounded to allow for better
understanding.
Made gravity act based on the surrounding squares so that  lower or
higher gravity air block can be added.
Created a doGravity Method that is only called if not grounded.
Created a doXFriction method that uses the higher of the 2 friction
values below it as its basis
Tweaked detection of edges of world so that no null-pointers occur.
removed the gravity variable

-- SQUARE CHANGES
created a free variable (change later, couldnt think of appropriate
name) that indicates whether it can be moved through
created a gravity variable so that gravity is based on the square
added new low gravity green square (added this to test.level)